requirements.txt
fastapi==0.116.1
pydantic==2.11.7
pymongo==4.13.2
python-dotenv==1.1.1
uvicorn==0.35.0
